Extract function 'atPosition'
